commit after add landing page and navigator
