+++ .expo/packager-info.json
@@ -2,8 +2,8 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 34685,
+  "packagerPid": 39960,
   "expoServerNgrokUrl": "https://xg-zc6.mmeng.google-vision-rn-demo.exp.direct",
   "packagerNgrokUrl": "https://packager.xg-zc6.mmeng.google-vision-rn-demo.exp.direct",
-  "ngrokPid": 34700
+  "ngrokPid": 39974
 }
