+++ package.json
@@ -9,12 +9,16 @@
   },
   "dependencies": {
     "expo": "^34.0.1",
+    "expo-camera": "^6.0.0",
     "expo-image-picker": "^6.0.0",
     "firebase": "^6.6.1",
     "react": "16.8.3",
     "react-dom": "^16.8.6",
     "react-native": "https://github.com/expo/react-native/archive/sdk-34.0.0.tar.gz",
+    "react-native-gesture-handler": "^1.4.1",
     "react-native-web": "^0.11.4",
+    "react-navigation": "^4.0.5",
+    "react-navigation-stack": "^1.7.3",
     "uuid": "^3.3.3"
   },
   "devDependencies": {
